src/requirements.txt
streamlit
google.generativeai
dotenv
os